pyproject.toml
[project]
name = "tinysim"
version = "0.0.1"
description = "small modular simulation environments"
authors = [
    {name = "Matthew Taylor", email = "matthew.taylor.andre@gmail.com"},
]
requires-python = ">=3.9"
classifiers = [
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Framework :: Jupyter :: JupyterLab :: 4",
    "Framework :: Jupyter :: JupyterLab :: Extensions",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
urls = {Homepage = "https://github.com/MatthewAndreTaylor/TinySim"}
keywords = ["jupyter", "simulation"]

dependencies = [
    "numpy",
    "anywidget",
    "jupyter-ui-poll"
]

[project.optional-dependencies]
warp = ["warp-lang==1.8.1"]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tinysim", "tinysim_warp"]
include-package-data = true

[tool.setuptools.package-data]
tinysim = ["**/*"]
tinysim_warp = ["**/*"]


